Create a mockup for a asynch software fetcher
This is a rough example of how one can create a fetcher for <some stuff>
like for example software upgrades where there are many consumers of the
same resource (when the resource is fetched) and the resource can be
successfully fetched, or the fetch can fail.

Currently the logic for retrying failed fetches will be put on the
client - but the fetching could be extended to take for example an
optional 'retry_strategy' and retries for the fetching can be moved in
to the fetcher.

Additionally the fetcher should probably not run callbacks synchronously
to avoid blowing up the stack. This could be done by injecting an
'executor' or some other way to manage jobs and place callbacks in a
queue outside of the fetcher to be completed later.